backend/pyproject.toml
[project]
name = "api-de-tarefas"
version = "0.1.0"
description = ""
authors = [
    {name = "Igor Augusto",email = "igoraf.dev33@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.121.3,<0.122.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.0.1,<4.1.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)"
]

[tool.poetry]
packages = [{include = "api_de_tarefas", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
